REIS-197 fix falling man
