+++ app/component/RouteScheduleContainer.js
@@ -14,8 +14,6 @@ import Loading from './Loading';
 import Icon from './Icon';
 import { RealtimeStateType } from '../constants';
 import TimetableWeekViewPdf from './TimetableWeekViewPdf';
-import TimetableWeekViewPdfButton from './TimetableWeekViewPdfButton';
-import RouteStopListContainer from './RouteStopListContainer';
 
 const DATE_FORMAT = 'YYYYMMDD';
 
@@ -27,12 +25,36 @@ const isTripCanceled = trip =>
     .map(x => x[1])
     .every(st => st.realtimeState === RealtimeStateType.Canceled);
 
+const RouteTypeLine = color => {
+  switch (color) {
+    case 'ff711d':
+      return 'route-type-train';
+    case 'de2c42':
+      return 'route-type-city-line-bus';
+    case '016e12':
+      return 'route-type-city-line-tram';
+    case '1ccc48':
+      return 'route-type-city-line-trolley';
+    case 'bd4819':
+      return 'route-type-city-line-commercial';
+    case '3bb5db':
+      return 'route-type-regional-line-pso';
+    case '094f82':
+      return 'route-type-regional-line-commercial';
+    case '660000':
+      return 'route-type-long-distance-bus-line';
+    case '8bb4c5':
+      return 'route-type-ferry';
+    default:
+      return '';
+  }
+}
+
 class RouteScheduleContainer extends Component {
   static propTypes = {
     pattern: PropTypes.object.isRequired,
     relay: PropTypes.object.isRequired,
     serviceDay: PropTypes.string.isRequired,
-    patternForPdf: PropTypes.object,
   };
 
   static contextTypes = {
@@ -168,6 +190,37 @@ class RouteScheduleContainer extends Component {
     window.print();
   };
 
+  injectTranslationsForPdf = patterns => {
+    return patterns.map(pattern => {
+      // eslint-disable-next-line no-param-reassign
+      pattern.trip.serviceOperatorLabel = this.context.intl.formatMessage({
+        id: 'service-operator',
+        defaultMessage: 'Operator:',
+      });
+      // eslint-disable-next-line no-param-reassign
+      pattern.trip.serviceManagerLabel = this.context.intl.formatMessage({
+        id: 'service-manager',
+        defaultMessage: 'Manager:',
+      });
+      // eslint-disable-next-line no-param-reassign
+      pattern.trip.routeTypeLabel = this.context.intl.formatMessage({
+        id: RouteTypeLine(pattern.trip.route.color),
+        defaultMessage: 'Regional line (Commercial)',
+      });
+      // eslint-disable-next-line no-param-reassign
+      pattern.trip.routeValidFromLabel = this.context.intl.formatMessage({
+        id: 'timetable-valid-from',
+        defaultMessage: 'The timetable is valid from:',
+      });
+      // eslint-disable-next-line no-param-reassign
+      pattern.trip.routeValidTillLabel = this.context.intl.formatMessage({
+        id: 'timetable-valid-till',
+        defaultMessage: 'The timetable is valid until:',
+      });
+      return pattern;
+    });
+  };
+
   initState(props, isInitialState) {
     const state = {
       from: 0,
@@ -210,9 +263,12 @@ class RouteScheduleContainer extends Component {
             onDateChange={this.changeDate}
           />
           {showWeekView() && (
-            <TimetableWeekViewPdfButton
-              pattern={this.props.patternForPdf}
-            />
+              <TimetableWeekViewPdf
+                  {...this.props}
+                  patterns={this.injectTranslationsForPdf(
+                      this.props.pattern.route.patterns,
+                  )}
+              />
           )}
 
           {this.dateForPrinting()}
@@ -272,6 +328,67 @@ const connectedComponent = connectToStores(
             shortName
             color
             longName
+            patterns {
+                stops {
+                    name
+                }
+                trip {
+                    id
+                    gtfsId
+                    tripLongName
+                    wheelchairAccessible
+                    tripTimesWeekdaysGroups
+                    tripTimesValidTill
+                    route {
+                        color
+                        shortName
+                        longName
+                        agency {
+                            name
+                        }
+                        competentAuthority
+                    }
+                    stoptimesForWeek {
+                        parts
+                        weekdays
+                        calendarDatesByFirstStoptime {
+                            time
+                            calendarDateExceptions {
+                                exceptionType
+                                dates
+                            }
+                        }
+                        tripTimesByWeekdaysList {
+                            tripTimeByStopNameList {
+                                stopName
+                                differentDeparture
+                                tripTimeShort {
+                                    headsign
+                                    realtimeState
+                                    scheduledArrival
+                                    scheduledDeparture
+                                    serviceDay
+                                    pickupType
+                                    dropoffType
+                                }
+                            }
+                        }
+                    }
+                }
+                patternTimetable (stopId: $stopId) {
+                  validity {
+                    validFrom
+                    validTill
+                  }
+                  weekdays
+                  trip {
+                    id
+                    departureStoptime (serviceDate: $serviceDate) {
+                      scheduledDeparture
+                    }
+                  }
+                }
+            }
           }
           tripsForDate(serviceDay: $serviceDay) {
             id
@@ -289,13 +406,6 @@ const connectedComponent = connectToStores(
           }
         }
       `,
-      patternForPdf: () =>
-        Relay.QL`
-      fragment on Pattern {
-        code
-        ${TimetableWeekViewPdfButton.getFragment('pattern')}
-      }
-    `,
     },
   }),
   [],
